requirements.txt
python-telegram-bot==20.3
paho-mqtt==1.6.1
transformers==4.47.0
pyyaml==6.0
nest-asyncio==1.5.7
